+++ common/src/main/java/com/feed_the_beast/ftbquests/quest/task/Task.java
@@ -70,7 +70,7 @@ public abstract class Task extends QuestObject
 	}
 
 	@Override
-	public final int getParentID()
+	public final long getParentID()
 	{
 		return quest.id;
 	}
